backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.2.3
python-dotenv==0.19.0 